+++ app/src/main/java/com/digiventure/ventnote/feature/notes/components/dialog/FilterDialog.kt
@@ -1,0 +1,8 @@
+package com.digiventure.ventnote.feature.notes.components.dialog
+
+import androidx.compose.runtime.Composable
+
+@Composable
+fun FilterDialog() {
+    
+}
