+++ app/src/main/res/values/strings.xml
@@ -9,6 +9,7 @@
     <string name="close_nav_icon">close nav icon</string>
     <string name="drawer_nav_icon">drawer nav icon</string>
     <string name="search_nav_icon">search nav icon</string>
+    <string name="more_nav_icon">more nav icon</string>
     <string name="delete_nav_icon">delete nav icon</string>
     <string name="dropdown_nav_icon">dropdown nav icon</string>
     <string name="back_nav_icon">back nav icon</string>
